Netlink

+++ workspace/linux-support/src/seccomp/c/seccomp_data.rs
@@ -40,6 +40,6 @@ impl seccomp_data
 	#[allow(unused)]
 	pub(crate) fn allocate_zeroed() -> VariablySized<Self>
 	{
-		VariablySized::allocate_zeroed(seccomp_notif_sizes::get_listener_notification_sizes().seccomp_data)
+		VariablySized::allocate_zeroed(seccomp_notif_sizes::get_listener_notification_sizes().seccomp_data as usize)
 	}
 }
